Extract Tests::Formulae
Extract the final class from Test. There should be a bit of cleanup
and DRYing based on this but that can wait for a future PR.
